Updated BDD tags logic
