+++ setup.cfg
@@ -18,8 +18,5 @@ target-version = ['py37']
 
 [tool:pytest]
 markers =
-    unit: Unit test Marker
-    integration: Integration test Marker
-    system: System test Marker
-    api: API test Marker
-    ui: UI test Marker
+    unit: Unit tests Marker
+    presence_ignore: Example Marker for ignore
